refactor service tests
